refactor food modal content to a separate file
