+++ screens/main/log/meal/FoodSearchEngine.js
@@ -16,8 +16,8 @@ import SampleFoodDB from './SampleFoodDB.json';
 import carbohydrate from '../../../../resources/images/icons/carbohydrate.png';
 import energy from '../../../../resources/images/icons/energy.png';
 import fat from '../../../../resources/images/icons/fat.png';
-import ProgressBar from "../../../../components/progressbar";
 import Icon from 'react-native-vector-icons/dist/FontAwesome';
+import FoodModalContent from "./FoodModalContent";
 
 Icon.loadFont();
 
@@ -50,6 +50,7 @@ class FoodSearchEngineScreen extends React.Component {
                 this.timeout = setTimeout(() => {
                     // Fetch api and load the response here.
                     // For now this would be just simulating an async process.
+                    /*
                     setTimeout(() => {
                         this.setState({
                             isLoading: false,
@@ -57,9 +58,11 @@ class FoodSearchEngineScreen extends React.Component {
                         })
                     }, 1000) // simulate 1s for fetching request.
 
+
+                     */
                     // Format to fetch the data
-                    /*
-                    fetch('http://ip:port/food/search', {
+
+                    fetch('http://192.168.1.8:5000/food/search', {
                         method: "POST",
                         headers: {
                             "Content-Type": "application/json"
@@ -70,8 +73,12 @@ class FoodSearchEngineScreen extends React.Component {
                             foodResults: data.data,
                             isLoading: false
                         })
+                    }).catch(err => {
+                        this.setState({
+                            isLoading: false
+                        }, () => alert(err.message))
                     })
-                      */
+
                 }, 500); // 500ms delay before loading API.
             })
         }
@@ -189,66 +196,13 @@ function FoodResultList({foodList, navigation, route, type}) {
         <React.Fragment>
             <Modal visible={modalOpen} transparent={true}>
                 {selected &&
-                    <View style={modalStyles.root}>
-                        <TouchableOpacity style={modalStyles.overlay} onPress={handleClose} />
-                        <View style={modalStyles.paper}>
-                            <Image style={modalStyles.image} source={{uri: selected.imgUrl.url}}/>
-                            <View style={modalStyles.nutritionInfoContainer}>
-                                <ScrollView contentContainerStyle={{padding: 15}}>
-                                    <Text>{selected["food-name"][0].toUpperCase() + selected["food-name"].slice(1)}</Text>
-                                    <Text>{selected["household-measure"]}({selected["per-serving"]})</Text>
-                                    <Text>Nutritional Info per serving</Text>
-                                    <View style={modalStyles.nutrientRow}>
-                                        {renderNutritionText(selected.nutrients["energy"], "Energy")}
-                                        <ProgressBar progress="30%" useIndicatorLevel={true}
-                                                     containerStyle={{height: 15, width: '100%'}} />
-                                    </View>
-                                    <View style={modalStyles.nutrientRow}>
-                                        {renderNutritionText(selected.nutrients["carbohydrate"], "Carbohydrate")}
-                                        <ProgressBar progress="60%" useIndicatorLevel={true}
-                                                     containerStyle={{height: 15, width: '100%'}} />
-                                    </View>
-                                    <View style={modalStyles.nutrientRow}>
-                                        {renderNutritionText(selected.nutrients["protein"], "Protein")}
-                                        <ProgressBar progress="90%" useIndicatorLevel={true} reverse={true}
-                                                     containerStyle={{height: 15, width: '100%'}} />
-                                    </View>
-                                    <View style={modalStyles.nutrientRow}>
-                                        {renderNutritionText(selected.nutrients["total-fat"], "Total Fat")}
-                                        <ProgressBar progress="60%" useIndicatorLevel={true}
-                                                     containerStyle={{height: 15, width: '100%'}} />
-                                    </View>
-                                    <View style={modalStyles.nutrientRow}>
-                                        {renderNutritionText(selected.nutrients["saturated-fat"], "Saturated Fat")}
-                                        <ProgressBar progress="60%" useIndicatorLevel={true}
-                                                     containerStyle={{height: 15, width: '100%'}} />
-                                    </View>
-                                    <View style={modalStyles.nutrientRow}>
-                                        {renderNutritionText(selected.nutrients["dietary-fibre"], "Dietary Fibre")}
-                                        <ProgressBar progress="30%" useIndicatorLevel={true} reverse={true}
-                                                     containerStyle={{height: 15, width: '100%'}} />
-                                    </View>
-                                    <View style={modalStyles.nutrientRow}>
-                                        {renderNutritionText(selected.nutrients["cholesterol"], "Cholesterol")}
-                                        <ProgressBar progress="60%" useIndicatorLevel={true}
-                                                     containerStyle={{height: 15, width: '100%'}} />
-                                    </View>
-                                    {   selected.nutrients.sodium &&
-                                        <View style={modalStyles.nutrientRow}>
-                                            {renderNutritionText(selected.nutrients["sodium"], "Sodium")}
-                                            <ProgressBar progress="90%" useIndicatorLevel={true}
-                                                         containerStyle={{height: 15, width: '100%'}} />
-                                        </View>
-                                    }
-                                </ScrollView>
-                            </View>
+                       <FoodModalContent selected={selected} onClose={handleClose}>
                             <TouchableHighlight
-                                style={modalStyles.button}
+                                style={styles.button}
                                 underlayColor='#fff' onPress={navigateBackToCreateMealLog}>
-                                <Text style={modalStyles.buttonText}>Add</Text>
+                                <Text style={styles.buttonText}>Add</Text>
                             </TouchableHighlight>
-                        </View>
-                    </View>
+                       </FoodModalContent>
                 }
             </Modal>
             <FlatList style={listStyles.container} data={foodList}
@@ -260,7 +214,7 @@ function FoodResultList({foodList, navigation, route, type}) {
 function renderNutritionText({amount, unit}, nutrient) {
     return (
         <View style={{flexDirection: 'row', justifyContent: 'space-between'}}>
-            <Text>{nutrient}</Text>
+            <Text style={modalStyles.nutrientText}>{nutrient}</Text>
             {
                 unit === "N.A" ? <Text>{unit}</Text>
                     : <Text>{amount + " " + unit}</Text>
@@ -303,52 +257,6 @@ function getColorFromNutrient({rating}, nutrientName) {
         }
     }
 }
-
-const modalStyles = StyleSheet.create({
-    root: {
-        flex: 1,
-        alignItems: 'center',
-        justifyContent: 'center'
-    },
-    overlay: {
-        height: '100%',
-        width: '100%',
-        backgroundColor: 'rgba(0, 0, 0, 0.5)',
-        position: 'absolute'
-    },
-    paper: {
-        backgroundColor: 'white',
-        width: '80%',
-        height: '70%'
-    },
-    image: {
-        width: '100%',
-        height: '40%'
-    },
-    nutritionInfoContainer: {
-        height: '53.5%',
-    },
-    button: {
-        marginRight:40,
-        marginLeft:40,
-        marginTop:10,
-        paddingTop:20,
-        paddingBottom:20,
-        backgroundColor:'#68a0cf',
-        borderRadius:10,
-        borderWidth: 1,
-        borderColor: '#fff',
-        //transform: [{"translateY": '+42%'}]
-    },
-    buttonText: {
-        color:'#fff',
-        textAlign:'center',
-    },
-    nutrientRow: {
-        width: '100%',
-        paddingTop: 10
-    }
-});
 
 const listStyles = StyleSheet.create({
     container: {
@@ -450,7 +358,21 @@ const styles = StyleSheet.create({
     foodSearchResults: {
         flex: 1,
         alignItems: 'center'
-    }
+    },
+    button: {
+        width: '100%',
+        height: 55,
+        backgroundColor:'#aad326',
+        justifyContent: 'center',
+        alignSelf: 'center',
+        transform: [{"translateY": 27.5}] // Half of height
+    },
+    buttonText: {
+        color:'#000',
+        textAlign:'center',
+        fontSize: 22,
+        fontWeight: 'bold'
+    },
 })
 
 export default FoodSearchEngineScreen;
