+++ screens/main/log/meal/SampleFoodDB.json
@@ -107,7 +107,7 @@
             },
             "subcategory" : "All fast food outlets in alphabetical order",
             "imgUrl" : {
-                "url" : "http://4.bp.blogspot.com/_W-pG7tUmJk0/S-c6YFumqFI/AAAAAAAAB5c/fO50kG4hvUQ/s400/kfc_aus_roller.gif"
+                "url" : "https://suitabletech.com/images/HelpCenter/errors/Lenovo-Camera-Error.JPG"
             }
         },{
             "_id" : "5ec561a169eef8907d7921b3",
